Rename files to match project format.
