+++ components/graphs/ByCategory.tsx
